Reader: validate global archive header (#3)
Rather than discarding the first 8 bytes read from an `io.Reader` when
wrapping it with a `Reader`, ensure they match the ar file format's
global file header, and return an error if not. This makes ar's `Reader`
behave similarly to the `Reader`s in the `archive/*` modules in Go's
standard library.