Writer: ensure archive header is written first
Rather than expecting the user to call `WriteGlobalHeader` before
`WriteHeader` or `Write` (which unnecessarily exposes details of the ar
file format to users), instead ensure that the archive header is written
to the underlying `io.Writer` before any other data, and expect users to
call `Close` to finalise the archive (which ensures that a valid archive
header has been written). This follows a similar pattern to the archive
modules in the standard library, e.g. `archive/zip`.