Writer: require explicit output format
When creating a `Writer`, require the ar file format variant for the
generated archive to be explicitly specified. This allows us to be
stricter about how the `Writer` is used (e.g. by forbidding the creation
of a GNU-format archive with more than one string table).

Also, remove the `Writer` tests that ensure the output from `Writer` is
byte-equal with an archive file generated by some other ar tool - not
every ar implementation writes archives of a particular variant the same
way, and we're not aiming for byte-for-byte compatibility with an
existing tool. Replace these tests with end-to-end tests that make sure
`Writer` generates archives that can be read by a range of third-party
ar tools (GNU ar, llvm-ar, macOS ar).